requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
nebula3-python
llama_index.llms.ollama
tqdm